added testing, split files, user and adventures working on Postman
